refactor: separate signIn and signUp handlers
