+++ src/entities/AuthForm/types/index.ts
@@ -1,0 +1,6 @@
+import type * as Yup from 'yup';
+import { type validationSchemaSignUpWithEmail, type validationSchemaSignInWithEmail } from '../lib/validationSchema';
+
+export type SignUpWithEmailData = Yup.InferType<typeof validationSchemaSignUpWithEmail>;
+
+export type SignInWithEmailData = Yup.InferType<typeof validationSchemaSignInWithEmail>;
